xhr/access-control-and-redirects-async-same-origin.htm

<!DOCTYPE html>
<html>
  <head>
    <title>Tests that asynchronous XMLHttpRequests handle redirects according to the CORS standard.</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
  </head>
  <body>
    <script>
    function runTest(testName, path, credentials, expectSuccess) {
      return promise_test(test => {
          return new Promise(resolve => {
              const xhr = new XMLHttpRequest();
              xhr.withCredentials = credentials;
              xhr.open("GET", "resources/redirect.py?location=" + get_host_info().HTTP_REMOTE_ORIGIN + path, true);

              xhr.onload = test.step_func(function() {
                  assert_true(expectSuccess);
                  assert_equals(xhr.responseText, "PASS: Cross-domain access allowed.");
                  resolve();
              });
              xhr.onerror = test.step_func(function() {
                  assert_false(expectSuccess);
                  assert_equals(xhr.status, 0);
                  resolve();
              });
              xhr.send(null);
          });
      }, testName);
    }

    const withoutCredentials = false;
    const withCredentials = true;
    const succeeds = true;
    const fails = false;

    // Test simple same origin requests that receive cross origin redirects.

    // The redirect response passes the access check.
    runTest("Request without credentials is redirected to a cross-origin response with Access-Control-Allow-Origin=* (with star)",
            "/xhr/resources/access-control-basic-allow-star.py",
            withoutCredentials, succeeds);

    // The redirect response fails the access check because credentials were sent.
    runTest("Request with credentials is redirected to a cross-origin response with Access-Control-Allow-Origin=* (with star)",
            "/xhr/resources/access-control-basic-allow-star.py",
            withCredentials, fails);

    // The redirect response passes the access check.
    runTest("Request without credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin",
            "/xhr/resources/access-control-basic-allow.py",
            withoutCredentials, succeeds);

    // The redirect response passes the access check.
    runTest("Request with credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin",
            "/xhr/resources/access-control-basic-allow.py",
            withCredentials, succeeds);

    // forbidding credentials. The redirect response passes the access check.
    runTest("Request without credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin (no credentials)",
            "/xhr/resources/access-control-basic-allow-no-credentials.py",
            withoutCredentials, succeeds);

    // forbidding credentials. The redirect response fails the access check.
    runTest("Request with credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin (no credentials)",
            "/xhr/resources/access-control-basic-allow-no-credentials.py",
            withCredentials, fails);
    </script>
  </body>
</html>
